requirements.txt
accelerate
aiohttp
anyio
bitsandbytes
bs4
diffusers
transformers
faiss-cpu
fastapi
git-python
google-search-results
httptools
huggingface
huggingface-hub
json5
langchain
numexpr
llama_cpp_python
psutil
PyMuPDF
safetensors
selenium
sentence-transformers
sentencepiece
streamlit_chat
streamlit>=0.86.0
textblob
undetected-chromedriver
urllib3
virtualenv
wikipedia
datetime
asyncio
requests
bs4
Pillow
langchain
googletrans
torch
torchaudio
#torchaudio==2.0.1+cu117
torchvision
python-dotenv
lxml
ctransformers
omegaconf